Remove CroppedMetric from cropped_metric.py and implement it in transformed_metrics.py; add ResizeMetric class for enhanced resizing functionality with aspect ratio support.
